backend/requirements.txt
flask==3.1.0
flask-cors==6.0.1
requests==2.32.4
boto3==1.40.1
python-dotenv==1.0.1
youtube-transcript-api==1.2.2
gunicorn==21.2.0